+++ CHANGELOG.rst
@@ -2,7 +2,7 @@
 -------------------
 
 *	Fixed attribute name in ``AppLookupControl.vsqlfield``
-*	Remove debug ``print``s.
+*	Remove debug ``print``\s.
 
 
 0.50.1 (2025-09-10)