+++ CHANGELOG.rst
@@ -1,11 +1,15 @@
 HEAD (2025-??-??)
 -----------------
 
+*	Python 3.14 is required now, since we're using type hints that only work
+	with deferred annotations.
 *	Added class ``Permissions`` and attribute ``App.permissions``.
 *	Added attribute ``App.recordpage``.
 *	Added attribute ``Globals.user_record``.
 *	Added attributes ``App.filter_default``, ``App.sort_default`` and
 	``App.filter_owndata``.
+*	Added methods ``Record.fetch_child_records``, ``Record.count_child_records``
+	and ``Record.fetch_child_recordpage``.
 
 
 0.50.2 (2025-09-16)